requirements.txt
plotly==5.22.0
regex==2024.5.15
requests==2.31.0
sentence-transformers==2.2.2
torch==2.3.0
tqdm==4.66.4
transformers==4.41.0
umap-learn==0.5.6
scikit-learn>=1.0
notebook>=6.0
ipywidgets>=7.6
